+++ CHANGELOG.md
@@ -5,15 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.59.0] - Unpublished
+## [0.59.0] - 2024-05-02
 
 ### Added
 
 ### Fixed
 
+- [bug](https://github.com/FrankC01/pysui/issues/201) - Parameter passing fix for JSON RPC
 - [bug](https://github.com/FrankC01/pysui/issues/200) - Typo in cursor reference in sample code
 
 ### Changed
+
+- GetTx querynode includes richer object changes effects content.
 
 ### Removed